+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to Cranberry\CLI will be documented in this file.
 
 ## [Unreleased]
+### Added
+- `phpMinimumVersion` property to `Application\Mirror`
+
 ### Fixed
 - Switched formatted Shell::exec output to `Format\String`